Setup

+++ controllers/dietsController.js
@@ -4,22 +4,40 @@ import path from "path";
 import crypto from "crypto";
 import {ApiError} from "../utils/errors.js";
 import {checkPermissions} from "../utils/auth.js";
+import Meal from "../models/mealsModel.js";
 
 export const getDiets = async (req, res, next) => {
-    if(!checkPermissions(req.userInfo, process.env.ACCESS_USER)){
+    if(!await checkPermissions(req.userInfo, process.env.ACCESS_USER)){
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
+    const page = parseInt(req.query.page);
+    const pageSize = parseInt(req.query.pageSize);
+    const dietType = String(req.query.dietType);
     //CAUTION! Here will be also fetched the generic flexi diet! With this syntax we will get only fixed diets
     try {
-        const diets = await Diet.find({diet_type: 'Fixed'}).populate('exclusions').populate('tags_id');
+        const diets = await Diet.find(dietType ? {diet_type: dietType} : {}).populate('exclusions').populate('tags_id').skip((page - 1) * pageSize).limit(pageSize);
         if (diets) {
             const dietsImage = diets.map((diet) => {
+                if(diet.img.img_path === ''){
+                    return {...diet._doc, imageBuffer: null}
+                }
                 let dietPath = path.join('public', 'images', diet.img.img_path);
                 const data = fs.readFileSync(dietPath, {encoding: 'base64'})
                 return {...diet._doc, imageBuffer: data}
             })
+            const totalItems = await Diet.find({diet_type: 'Fixed'}).countDocuments()
             res.status(200);
-            res.json(dietsImage)
+            res.json({
+                diets: dietsImage,
+                paginationInfo: {
+                    totalItems,
+                    hasNextPage: pageSize * page < totalItems,
+                    hasPreviousPage: page > 1,
+                    nextPage: page + 1,
+                    previousPage: page - 1,
+                    lastPage: Math.ceil(totalItems / pageSize)
+                }
+            })
         }
 
     } catch (e) {
@@ -27,17 +45,20 @@ export const getDiets = async (req, res, next) => {
     }
 }
 export const getDiet = async (req, res, next) => {
-    if(!checkPermissions(req.userInfo, process.env.ACCESS_USER)){
+    if(!await checkPermissions(req.userInfo, process.env.ACCESS_USER)){
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const id = req.params.id;
+    let imgData = null;
     try {
         const diet = await Diet.findById(id).populate('exclusions').populate('tags_id');
         if (!diet) {
             return next(ApiError('Diet does not exist!', 404))
         }
         const dietImgPath = path.join('public', 'images', diet.img.img_path)
-        const imgData = fs.readFileSync(dietImgPath, {encoding: 'base64'});
+        if(diet.img.img_path !== ''){
+         imgData = fs.readFileSync(dietImgPath, {encoding: 'base64'});
+        }
         res.status(200)
         res.json({...diet._doc, imageBuffer: imgData});
 
@@ -46,21 +67,25 @@ export const getDiet = async (req, res, next) => {
     }
 }
 export const createDiet = async (req, res, next) => {
-    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+    if(!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const file = req.file;
+    let imgObj = {img_path: '', uri: ''};
     try {
         const dietData = JSON.parse(req.body.data)
         const existingDiet = await Diet.findOne({name: dietData.name});
         if (existingDiet) {
             return next(ApiError('Diet already exist!', 409))
         }
+        if(file){
         const fileBytes = fs.readFileSync(file.path);
-        const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
-        const imgObj = {
+        const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex');
+         imgObj = {
             img_path: file.originalname,
             uri: fileUri
+        }
+
         }
         const newDiet = new Diet({
             ...dietData,
@@ -79,18 +104,23 @@ export const createDiet = async (req, res, next) => {
     }
 }
 export const updateDiet = async (req, res, next) => {
-    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+    if(!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const id = req.params.id;
+    let fileObj = {img_path: '', uri: ''}
+    let fileUri = ''
     const file = req.file;
     try {
         const dietData = JSON.parse(req.body.data);
+        if(file){
         const fileBytes = fs.readFileSync(file.path);
-        const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
-        const fileObj = {
+        fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
+            fileObj = {
             img_path: file.originalname,
             uri: fileUri
+        }
+
         }
 
         const diet = await Diet.findByIdAndUpdate(id, {
@@ -103,11 +133,17 @@ export const updateDiet = async (req, res, next) => {
         //compare image uris
         if (fileUri !== diet.img.uri) {
             const dietPath = path.join('public', 'images', diet.img.img_path)
-            fs.unlink(dietPath, (err) => {
-                if (err) {
-                    return next(err);
+            if(diet.img.img_path !== ''){
+                const mealsWithSameImage = await Meal.find({"img.img_path": diet.img.img_path})
+                const dietsWithSameImage = await Diet.find({"img.img_path": diet.img.img_path})
+                if(mealsWithSameImage.length === 0 && dietsWithSameImage.length === 0){
+                    fs.unlink(dietPath, (err) => {
+                        if (err) {
+                            return next(err);
+                        }
+                    })
                 }
-            })
+            }
         }
         res.status(200);
         res.json({message: 'Diet updated!'})
@@ -118,7 +154,7 @@ export const updateDiet = async (req, res, next) => {
 }
 //action only for admin!
 export const deleteDiet = async (req, res, next) => {
-    if(!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
+    if(!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)){
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     //meals are not affected
@@ -130,11 +166,17 @@ export const deleteDiet = async (req, res, next) => {
         }
         const mealPath = path.join('public', 'images', deletedDiet.img.img_path)
         console.log(mealPath)
-        fs.unlink(mealPath, (err) => {
-            if (err) {
-                throw (err)
+        if(deletedDiet.img.img_path !== ''){
+            const mealsWithSameImage = await Meal.find({"img.img_path": deletedDiet.img.img_path})
+            const dietsWithSameImage = await Diet.find({"img.img_path": deletedDiet.img.img_path})
+            if(mealsWithSameImage.length === 0 && dietsWithSameImage.length === 0){
+                fs.unlink(mealPath, (err) => {
+                    if (err) {
+                        return next(err);
+                    }
+                })
             }
-        })
+        }
         res.status(200);
         res.json({message: 'diet deleted!'})
     } catch (e) {
