+++ controllers/mealsController.js
@@ -1,4 +1,5 @@
 import Meal from '../models/mealsModel.js'
+import Diet from '../models/dietsModel.js'
 import fs from "fs";
 import path from "path";
 import crypto from "crypto";
@@ -6,7 +7,7 @@ import {ApiError} from "../utils/errors.js";
 import {checkPermissions} from "../utils/auth.js";
 
 export const getMeals = async (req, res, next) => {
-    if (!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
+    if (!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const page = parseInt(req.query.page);
@@ -15,18 +16,21 @@ export const getMeals = async (req, res, next) => {
         const meals = await Meal.find({}).populate('exclusions').populate('tags_id').skip((page - 1) * pageSize).limit(pageSize)
         if (meals) {
             const mealsImage = meals.map((meal) => {
-                let mealPath = path.join('public', 'images', meal.img_path);
+                if(meal.img.img_path === ''){
+                    return {...meal._doc, imageBuffer: null}
+                }
+                let mealPath = path.join('public', 'images', meal.img.img_path);
                 const data = fs.readFileSync(mealPath, {encoding: 'base64'})
                 return {...meal._doc, imageBuffer: data}
             })
             const totalItems = await Meal.find({}).countDocuments()
             res.status(200)
             res.json({
-                mealsImage,
+                meals: mealsImage,
                 paginationInfo: {
                     totalItems,
                     hasNextPage: pageSize * page < totalItems,
-                    haPreviousPage: page > 1,
+                    hasPreviousPage: page > 1,
                     nextPage: page + 1,
                     previousPage: page - 1,
                     lastPage: Math.ceil(totalItems / pageSize)
@@ -39,34 +43,41 @@ export const getMeals = async (req, res, next) => {
 
 }
 export const getMealById = async (req, res, next) => {
-    if (!checkPermissions(req.userInfo, process.env.ACCESS_USER)) {
+    if (!await checkPermissions(req.userInfo, process.env.ACCESS_USER)) {
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const id = req.params.id;
     try {
-        const meal = await Meal.findById(id);
+        const meal = await Meal.findById(id).populate('exclusions').populate('tags_id');
         if (!meal) {
             return next(ApiError('Meal does not exist', 404))
         }
+        let mealPath = path.join('public', 'images', meal.img.img_path);
+        const data = fs.readFileSync(mealPath, {encoding: 'base64'})
+
         res.status(200)
-        res.json(meal)
+        res.json({...meal._doc, imageBuffer: data})
     } catch (e) {
         next(e);
     }
 
 }
 export const createMeal = async (req, res, next) => {
-    if (!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
+    if (!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     try {
         const mealData = JSON.parse(req.body.data);
+        let fileObj = {img_path: '', uri: ''}
         const file = req.file;
-        const fileBytes = fs.readFileSync(file.path);
-        const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
-        const fileObj = {
-            img_path: file.originalname,
-            uri: fileUri
+        console.log(file)
+        if (file) {
+            const fileBytes = fs.readFileSync(file.path);
+            const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
+            fileObj = {
+                img_path: file.originalname,
+                uri: fileUri
+            }
         }
         const meal = new Meal({
             ...mealData,
@@ -86,18 +97,22 @@ export const createMeal = async (req, res, next) => {
 
 }
 export const updateMeal = async (req, res, next) => {
-    if (!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
+    if (!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const file = req.file
+    let fileObj = {img_path: '', uri: ''}
+    let fileUri = ''
     const id = req.params.id
     try {
         const mealData = JSON.parse(req.body.data)
+        if(file){
         const fileBytes = fs.readFileSync(file.path);
-        const fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
-        const fileObj = {
+        fileUri = crypto.createHash('sha1').update(fileBytes).digest('hex')
+        fileObj = {
             img_path: file.originalname,
             uri: fileUri
+        }
         }
         const existingMeal = await Meal.findByIdAndUpdate(id, {
             ...mealData,
@@ -108,11 +123,17 @@ export const updateMeal = async (req, res, next) => {
         if (fileUri !== existingMeal.img.uri) {
             const mealPath = path.join('public', 'images', existingMeal.img.img_path)
             console.log(mealPath)
-            fs.unlink(mealPath, (err) => {
-                if (err) {
-                    return next(err);
+            if(existingMeal.img.img_path !== ''){
+                const mealsWithSameImage = await Meal.find({"img.img_path": existingMeal.img.img_path})
+                const dietsWithSameImage = await Diet.find({"img.img_path": existingMeal.img.img_path})
+                if(mealsWithSameImage.length === 0 && dietsWithSameImage.length === 0){
+                    fs.unlink(mealPath, (err) => {
+                        if (err) {
+                            return next(err);
+                        }
+                    })
                 }
-            })
+            }
         }
         res.status(200)
         res.json({message: 'Meal updated'})
@@ -122,7 +143,7 @@ export const updateMeal = async (req, res, next) => {
 
 }
 export const deleteMeal = async (req, res, next) => {
-    if (!checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
+    if (!await checkPermissions(req.userInfo, process.env.ACCESS_DIETETICIAN)) {
         return next(ApiError("You're not authorized to perform this action!", 401))
     }
     const id = req.params.id
@@ -133,11 +154,17 @@ export const deleteMeal = async (req, res, next) => {
         }
         const mealPath = path.join('public', 'images', deletedMeal.img.img_path)
         console.log(mealPath)
-        fs.unlink(mealPath, (err) => {
-            if (err) {
-                return next(err);
+        if(deletedMeal.img.img_path !== ''){
+            const mealsWithSameImage = await Meal.find({"img.img_path": deletedMeal.img.img_path})
+            const dietsWithSameImage = await Diet.find({"img.img_path": deletedMeal.img.img_path})
+            if(mealsWithSameImage.length === 0 && dietsWithSameImage.length === 0){
+                fs.unlink(mealPath, (err) => {
+                    if (err) {
+                        return next(err);
+                    }
+                })
             }
-        })
+        }
         res.status(200);
         res.json({message: 'Meal deleted!'})
